more refactoring to help make following the pathing descion easier
