+++ snakes/go/pathy-snake/coord-funcs.go
