+++ snakes/go/pathy-snake/gamestate-funcs.go
